+++ corebc-core/src/macros/ethers_crate.rs
@@ -74,9 +74,9 @@ impl ProjectEnvironment {
     /// project.
     ///
     /// The names will be:
-    /// - `ethers::*` if `ethers` is a dependency for all crates;
+    /// - `corebc::*` if `ethers` is a dependency for all crates;
     /// - for each `crate`:
-    ///   - `ethers_<crate>` if it is a dependency, otherwise `ethers::<crate>`.
+    ///   - `ethers_<crate>` if it is a dependency, otherwise `corebc::<crate>`.
     #[inline]
     pub fn determine_ethers_crates(&self) -> Option<CrateNames> {
         let lock_file = self.manifest_dir.join("Cargo.lock");
@@ -100,7 +100,7 @@ impl ProjectEnvironment {
         // return ethers_* if the root package is an internal ethers crate since `ethers` is not
         // available
         if pkg.name.parse::<CorebcCrate>().is_ok() || pkg.name == "ethers" {
-            return Some(CorebcCrate::path_names().collect())
+            return Some(CorebcCrate::path_names().collect());
         }
 
         let mut names: CrateNames = CorebcCrate::ethers_path_names().collect();
@@ -108,7 +108,7 @@ impl ProjectEnvironment {
             let name = dep.name.as_str();
             if name.starts_with("ethers") {
                 if name == "ethers" {
-                    return None
+                    return None;
                 } else if let Ok(dep) = name.parse::<CorebcCrate>() {
                     names.insert(dep, dep.path_name());
                 }
@@ -130,12 +130,12 @@ impl ProjectEnvironment {
     /// [ref]: https://doc.rust-lang.org/cargo/reference/environment-variables.html#environment-variables-cargo-sets-for-crates
     #[inline]
     pub fn is_crate_root(&self) -> bool {
-        env::var_os("CARGO_TARGET_TMPDIR").is_none() &&
-            self.manifest_dir.components().all(|c| {
+        env::var_os("CARGO_TARGET_TMPDIR").is_none()
+            && self.manifest_dir.components().all(|c| {
                 let s = c.as_os_str();
                 s != "examples" && s != "benches"
-            }) &&
-            !self.is_crate_name_in_dirs()
+            })
+            && !self.is_crate_name_in_dirs()
     }
 
     /// Returns whether `crate_name` is the name of a file or directory in the first level of
@@ -274,11 +274,11 @@ impl CorebcCrate {
         }
     }
 
-    /// "::ethers::`<self in ethers>`"
+    /// "::corebc::`<self in ethers>`"
     #[inline]
     pub const fn ethers_path_name(self) -> &'static str {
         match self {
-            // re-exported in ethers::contract
+            // re-exported in corebc::contract
             Self::CorebcContractAbigen => "::corebc::contract", // partially
             Self::CorebcContractDerive => "::corebc::contract",
 
@@ -309,7 +309,7 @@ impl CorebcCrate {
         Self::iter().map(|x| (x, x.path_name()))
     }
 
-    /// `<ethers::*>`
+    /// `<corebc::*>`
     #[inline]
     pub fn ethers_path_names() -> impl Iterator<Item = (Self, &'static str)> {
         Self::iter().map(|x| (x, x.ethers_path_name()))
@@ -328,7 +328,7 @@ impl CorebcCrate {
 fn file_stem_eq<T: AsRef<Path>, U: AsRef<str>>(path: T, s: U) -> bool {
     if let Some(stem) = path.as_ref().file_stem() {
         if let Some(stem) = stem.to_str() {
-            return stem == s.as_ref()
+            return stem == s.as_ref();
         }
     }
     false