+++ corebc-signers/src/wallet/yubi.rs
@@ -76,7 +76,7 @@ impl From<YubiSigner<Secp256k1>> for Wallet<YubiSigner<Secp256k1>> {
 mod tests {
     use super::*;
     use crate::Signer;
-    use std::str::FromStr;
+    use corebc_core::types::Address;
 
     #[tokio::test]
     async fn from_key() {
@@ -98,7 +98,7 @@ mod tests {
         assert_eq!(sig.recover(msg).unwrap(), wallet.address());
         assert_eq!(
             wallet.address(),
-            Address::from_str("2DE2C386082Cff9b28D62E60983856CE1139eC49").unwrap()
+            Address::from_str("00002DE2C386082Cff9b28D62E60983856CE1139eC49").unwrap()
         );
     }